TSK-1349: moved data to mock-store, extend workbasket information tests
