+++ web/src/app/administration/components/workbasket-information/workbasket-information.component.spec.ts
@@ -1,7 +1,7 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { WorkbasketInformationComponent } from './workbasket-information.component';
 import { Component, DebugElement, Input } from '@angular/core';
-import { Actions, NgxsModule, Store } from '@ngxs/store';
+import { Actions, NgxsModule, ofActionDispatched, Store } from '@ngxs/store';
 import { Observable, of } from 'rxjs';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { ICONTYPES } from '../../../shared/models/icon-types';
@@ -24,8 +24,13 @@ import { ClassificationCategoriesService } from '../../../shared/services/classi
 import { ACTION } from '../../../shared/models/action';
 import { TypeaheadModule } from 'ngx-bootstrap';
 import { TypeAheadComponent } from '../../../shared/components/type-ahead/type-ahead.component';
-import { Workbasket } from '../../../shared/models/workbasket';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
+import { MarkWorkbasketForDeletion, UpdateWorkbasket } from '../../../shared/store/workbasket-store/workbasket.actions';
+import {
+  selectedWorkbasket,
+  engineConfigurationMock,
+  workbasketReadState
+} from '../../../shared/store/mock-data/mock-store';
 
 @Component({ selector: 'taskana-shared-spinner', template: '' })
 class SpinnerStub {
@@ -54,218 +59,6 @@ const workbasketServiceMock = jest.fn().mockImplementation(
     createWorkbasket: jest.fn().mockReturnValue(of({ ...selectedWorkbasket }))
   })
 );
-
-export const engineConfigurationMock = {
-  customisation: {
-    EN: {
-      workbaskets: {
-        information: {
-          owner: {
-            lookupField: true
-          },
-          custom1: {
-            field: 'Customized field 1 title',
-            visible: true
-          },
-          custom3: {
-            field: '',
-            visible: false
-          }
-        },
-        'access-items': {
-          accessId: {
-            lookupField: true
-          },
-          custom3: {
-            field: '',
-            visible: false
-          },
-          custom9: {
-            field: 'Some custom field',
-            visible: true
-          },
-          custom10: {
-            field: '',
-            visible: false
-          },
-          custom11: {
-            field: '',
-            visible: false
-          },
-          custom12: {
-            field: '',
-            visible: false
-          }
-        }
-      }
-    }
-  },
-  language: 'EN'
-};
-
-const selectedWorkbasket: Workbasket = {
-  workbasketId: 'WBI:000000000000000000000000000000000902',
-  key: 'sOrt003',
-  name: 'bAsxet2',
-  domain: 'DOMAIN_A',
-  type: ICONTYPES.TOPIC,
-  description: 'Lorem ipsum dolor sit amet.',
-  owner: 'Max',
-  custom1: '',
-  custom2: '',
-  custom3: '',
-  custom4: '',
-  orgLevel1: '',
-  orgLevel2: '',
-  orgLevel3: '',
-  orgLevel4: '',
-  markedForDeletion: false,
-  created: '2020-08-18T09:14:41.353Z',
-  modified: '2020-08-18T09:14:41.353Z',
-  _links: {
-    self: {
-      href: 'http://localhost:8080/taskana/api/v1/workbaskets/WBI:000000000000000000000000000000000902'
-    },
-    distributionTargets: {
-      href:
-        'http://localhost:8080/taskana/api/v1/workbaskets/WBI:000000000000000000000000000000000902/distribution-targets'
-    },
-    accessItems: {
-      href:
-        'http://localhost:8080/taskana/api/v1/workbaskets/WBI:000000000000000000000000000000000902/workbasketAccessItems'
-    },
-    allWorkbaskets: {
-      href: 'http://localhost:8080/taskana/api/v1/workbaskets'
-    },
-    removeDistributionTargets: {
-      href:
-        'http://localhost:8080/taskana/api/v1/workbaskets/WBI:000000000000000000000000000000000902/distribution-targets'
-    }
-  }
-};
-
-export const workbasketReadState = {
-  selectedWorkbasket,
-  paginatedWorkbasketsSummary: {
-    _links: {
-      self: {
-        href:
-          'http://localhost:8080/taskana/api/v1/workbaskets/?sort-by=name&order=asc&domain=DOMAIN_A&page=3&page-size=8'
-      },
-      first: {
-        href:
-          'http://localhost:8080/taskana/api/v1/workbaskets/?sort-by=name&order=asc&domain=DOMAIN_A&page-size=8&page=1'
-      },
-      last: {
-        href:
-          'http://localhost:8080/taskana/api/v1/workbaskets/?sort-by=name&order=asc&domain=DOMAIN_A&page-size=8&page=3'
-      },
-      prev: {
-        href:
-          'http://localhost:8080/taskana/api/v1/workbaskets/?sort-by=name&order=asc&domain=DOMAIN_A&page-size=8&page=2'
-      }
-    },
-    workbaskets: [
-      {
-        workbasketId: 'WBI:100000000000000000000000000000000008',
-        key: 'USER-2-1',
-        name: 'PPK User 1 KSC 2',
-        domain: 'DOMAIN_A',
-        type: 'PERSONAL',
-        description: 'PPK User 1 KSC 2',
-        owner: '',
-        custom1: '',
-        custom2: '',
-        custom3: '',
-        custom4: '',
-        orgLevel1: '',
-        orgLevel2: '',
-        orgLevel3: '',
-        orgLevel4: '',
-        markedForDeletion: false
-      },
-      {
-        workbasketId: 'WBI:100000000000000000000000000000000007',
-        key: 'USER-1-2',
-        name: 'PPK User 2 KSC 1',
-        domain: 'DOMAIN_A',
-        type: 'PERSONAL',
-        description: 'PPK User 2 KSC 1',
-        owner: 'Peter Maier',
-        custom1: 'custom1',
-        custom2: 'custom2',
-        custom3: 'custom3',
-        custom4: 'custom4',
-        orgLevel1: 'versicherung',
-        orgLevel2: 'abteilung',
-        orgLevel3: 'projekt',
-        orgLevel4: 'team',
-        markedForDeletion: false
-      },
-      {
-        workbasketId: 'WBI:100000000000000000000000000000000009',
-        key: 'USER-2-2',
-        name: 'PPK User 2 KSC 2',
-        domain: 'DOMAIN_A',
-        type: 'PERSONAL',
-        description: 'PPK User 2 KSC 2',
-        owner: '',
-        custom1: '',
-        custom2: '',
-        custom3: '',
-        custom4: '',
-        orgLevel1: '',
-        orgLevel2: '',
-        orgLevel3: '',
-        orgLevel4: '',
-        markedForDeletion: false
-      },
-      {
-        workbasketId: 'WBI:100000000000000000000000000000000010',
-        key: 'TPK_VIP',
-        name: 'Themenpostkorb VIP',
-        domain: 'DOMAIN_A',
-        type: 'TOPIC',
-        description: 'Themenpostkorb VIP',
-        owner: '',
-        custom1: '',
-        custom2: '',
-        custom3: '',
-        custom4: '',
-        orgLevel1: '',
-        orgLevel2: '',
-        orgLevel3: '',
-        orgLevel4: '',
-        markedForDeletion: false
-      },
-      {
-        workbasketId: 'WBI:100000000000000000000000000000000016',
-        key: 'TPK_VIP_2',
-        name: 'Themenpostkorb VIP 2',
-        domain: 'DOMAIN_A',
-        type: 'TOPIC',
-        description: 'Themenpostkorb VIP',
-        owner: '',
-        custom1: '',
-        custom2: '',
-        custom3: '',
-        custom4: '',
-        orgLevel1: '',
-        orgLevel2: '',
-        orgLevel3: '',
-        orgLevel4: '',
-        markedForDeletion: false
-      }
-    ],
-    page: {
-      size: 8,
-      totalElements: 21,
-      totalPages: 3,
-      number: 3
-    }
-  },
-  action: ACTION.READ
-};
 
 const isFieldValidFn = jest.fn().mockReturnValue(true);
 const validateFormInformationFn = jest.fn().mockImplementation((): Promise<any> => Promise.resolve(true));
@@ -349,8 +142,10 @@ describe('WorkbasketInformationComponent', () => {
     expect(component).toBeTruthy();
   });
 
-  //HTML Tests
-  it('', () => {});
+  it('should display custom fields correctly', () => {
+    const customFields = debugElement.nativeElement.getElementsByClassName('custom-fields');
+    expect(customFields.length).toBe(3); //mock data has custom1->4 but engineConfig disables custom3 -> [1,2,4]
+  });
 
   it('should create clone of workbasket when workbasket value changes', () => {
     component.action = ACTION.READ;
@@ -391,21 +186,28 @@ describe('WorkbasketInformationComponent', () => {
     expect(component.workbasket).toMatchObject(component.workbasketClone);
   });
 
-  it('onSave', () => {
-    component.onSave();
-  });
-
-  it('postNewWOrkbasket', () => {
+  it('should save workbasket if workbasketId there', () => {
+    component.workbasket = { ...selectedWorkbasket };
+    component.workbasket.workbasketId = '1';
     component.action = ACTION.COPY;
-    component.postNewWorkbasket();
-  });
-
-  it('onRemoveConfirmed', () => {
-    component.onRemoveConfirmed();
-  });
-
-  it('onSave no id', () => {
-    delete component.workbasket.workbasketId;
+    let actionDispatched = false;
+    actions$.pipe(ofActionDispatched(UpdateWorkbasket)).subscribe(() => (actionDispatched = true));
     component.onSave();
+    expect(actionDispatched).toBe(true);
+    expect(component.workbasketClone).toMatchObject(component.workbasket);
+  });
+
+  it('should dispatch MarkWorkbasketforDeletion action when onRemoveConfirmed is called', () => {
+    let actionDispatched = false;
+    actions$.pipe(ofActionDispatched(MarkWorkbasketForDeletion)).subscribe(() => (actionDispatched = true));
+    component.onRemoveConfirmed();
+    expect(actionDispatched).toBe(true);
+  });
+
+  it('should create new workbasket if workbasketId is undefined', () => {
+    component.workbasket.workbasketId = undefined;
+    const postNewWorkbasketSpy = jest.spyOn(component, 'postNewWorkbasket');
+    component.onSave();
+    expect(postNewWorkbasketSpy).toHaveBeenCalled();
   });
 });
